+++ seeds/1-users.js
@@ -1,15 +1,14 @@
 // Enable ECMAScript module loader
-require = require("esm")(module); // eslint-disable-line
+require = require('esm')(module); // eslint-disable-line
 
 const users = require('../app/models/users');
 const { createMockUser } = require('../test/utils/mock-factory');
 
-exports.seed = async function (knex) {
-
+exports.seed = async function(knex) {
   const usersCount = await users.count();
   if (usersCount > 0) {
     // eslint-disable-next-line
-    console.log('There are USERS in the database already, bypass seeding...'); 
+    console.log('There are USERS in the database already, bypass seeding...');
     return;
   }
